requirements.txt
# =============================================================================
# Healthcare Hackathon — Python Dependencies
# =============================================================================

# --- API Framework ---
fastapi==0.115.6
uvicorn[standard]==0.32.1

# --- Data Processing ---
pandas==2.2.3
numpy==2.1.3

# --- Machine Learning ---
scikit-learn==1.5.2

# --- Visualization ---
matplotlib==3.9.3
seaborn==0.13.2

# --- Utilities ---
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2          # async HTTP client (used by FastAPI test client too)

# --- Health Data / FHIR ---
fhirclient==4.2.0      # FHIR R4 Python client

# --- Notebooks ---
jupyter==1.1.1
ipykernel==6.29.5

# --- Dev / Testing (optional but useful at a hackathon) ---
pytest==8.3.4
pytest-asyncio==0.24.0
